# openapi.yaml
openapi: 3.0.3
info:
  title: Revvo Smart Price API
  description: |
    API do sistema Revvo Smart Price — plataforma de gestão de terminais de consulta de preços e exibição de mídia em lojas.

    ## Autenticação

    A API possui dois mecanismos de autenticação:

    ### 1. Web Session (Cookie JWT)
    Utilizada pelo painel administrativo. O login via `POST /api/auth/login` retorna um cookie `revvo_session` (httpOnly, SameSite=lax) com validade de 8 horas.
    Todas as rotas administrativas exigem esse cookie.

    ### 2. Terminal Bearer Token
    Utilizada pelos dispositivos físicos (terminais). O terminal obtém um JWT via `POST /api/terminal/activate` (primeira vez) ou `POST /api/terminal/auth` (reautenticação).
    O token deve ser enviado no header `Authorization: Bearer <token>` e tem validade de 30 dias.

    ## Perfis de acesso

    - **Master** (`revvo-master`): acesso irrestrito a todas as empresas.
    - **Usuário comum**: acesso restrito à própria empresa.
  version: 1.0.0
  contact:
    name: Revvo Smart Price
servers:
  - url: https://revvo-smart-price-backend.vercel.app
    description: Produção
  - url: http://localhost:3000
    description: Local
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: revvo_session
      description: Cookie JWT gerado pelo login web. Válido por 8 horas.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT gerado na ativação do terminal. Válido por 30 dias.
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
      required: [message]
    Company:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Santo Antônio Supermercados
        legalName:
          type: string
          example: Santo Antônio Comercial Ltda
        cnpj:
          type: string
          example: "12.345.678/0001-99"
        email:
          type: string
          format: email
        slug:
          type: string
          example: santo-antonio
        status:
          type: string
          enum: [ATIVO, INATIVO]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        isActive:
          type: boolean
        companyId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        company:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ean:
          type: string
          example: "7891000315507"
        name:
          type: string
          example: Leite Integral 1L
        imageUrl:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        price:
          type: number
          format: decimal
          example: 4.99
        offerPrice:
          type: number
          nullable: true
          example: 3.99
        clubPrice:
          type: number
          nullable: true
          example: 3.49
        stock:
          type: integer
          nullable: true
        isActive:
          type: boolean
        externalId:
          type: string
          nullable: true
        storeId:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
...
    PaginatedProducts:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer
security:
  - cookieAuth: []
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login de usuário web
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login realizado com sucesso. Cookie `revvo_session` é definido.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      email:
                        type: string
                      companyId:
                        type: string
                      company:
                        type: object
        '400':
          description: E-mail e senha são obrigatórios
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: E-mail ou senha inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Usuário ou empresa inativos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/companies:
    get:
      tags: [Companies]
      summary: Listar empresas
      description: Retorna todas as empresas. Sem restrição de tenant (sem cookie obrigatório nos checks internos).
      responses:
        '200':
          description: Lista de empresas
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/Company'
                    - type: object
                      properties:
                        _count:
                          type: object
                          properties:
                            terminals:
                              type: integer
                            users:
                              type: integer
    post:
      tags: [Companies]
      summary: Criar empresa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, legalName, cnpj, email, slug]
              properties:
                name:
                  type: string
                legalName:
                  type: string
                cnpj:
                  type: string
                email:
                  type: string
                  format: email
                slug:
                  type: string
                status:
                  type: string
                  enum: [ATIVO, INATIVO]
                  default: ATIVO
      responses:
        '201':
          description: Empresa criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '400':
          description: Campos obrigatórios ausentes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: CNPJ, e-mail ou slug já cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/companies/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Companies]
      summary: Buscar empresa por ID
      responses:
        '200':
          description: Dados da empresa
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Company'
                  - type: object
                    properties:
                      users:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
                      terminals:
                        type: array
                        items:
                          type: object
        '404':
          description: Empresa não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: [Companies]
      summary: Editar empresa
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                legalName:
                  type: string
                cnpj:
                  type: string
                email:
                  type: string
                slug:
                  type: string
                status:
                  type: string
                  enum: [ATIVO, INATIVO]
      responses:
        '200':
          description: Empresa atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '409':
          description: CNPJ, e-mail ou slug duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Companies]
      summary: Remover empresa
      responses:
        '200':
          description: Empresa removida
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  # ─────────────────────────────────────────────
  # USERS
  # ─────────────────────────────────────────────
...
  /api/stores/{id}/products:
    parameters:
      - name: id
        in: path
        required: true
        description: ID da loja
        schema:
          type: string
          format: uuid
    get:
      tags: [Products]
      summary: Listar produtos da loja
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: q
          in: query
          description: Busca por nome ou EAN
          schema:
            type: string
      responses:
        '200':
          description: Produtos paginados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProducts'
    post:
      tags: [Products]
      summary: Criar ou atualizar produto (upsert por storeId+ean)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ean, name, price]
              properties:
                ean:
                  type: string
                name:
                  type: string
                price:
                  type: number
                offerPrice:
                  type: number
                  nullable: true
                clubPrice:
                  type: number
                  nullable: true
                stock:
                  type: integer
                  nullable: true
                imageUrl:
                  type: string
                  nullable: true
                isActive:
                  type: boolean
                  default: true
                externalId:
                  type: string
                  nullable: true
                hash:
                  type: string
                  nullable: true
                description:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Produto atualizado (upsert)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '201':
          description: Produto criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /api/stores/{id}/products/{productId}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID da loja
        schema:
          type: string
          format: uuid
      - name: productId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [Products]
      summary: Editar produto
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ean:
                  type: string
                name:
                  type: string
                price:
                  type: number
                offerPrice:
                  type: number
                  nullable: true
                clubPrice:
                  type: number
                  nullable: true
                stock:
                  type: integer
                  nullable: true
                imageUrl:
                  type: string
                  nullable: true
                isActive:
                  type: boolean
                description:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Produto atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Products]
      summary: Remover produto
      responses:
        '200':
          description: Produto removido
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  # ─────────────────────────────────────────────
  # TERMINALS
  # ─────────────────────────────────────────────
...